refactored some of the ballooning format files to be logic-separated
