+++ tests/test_ctramp_student_category.py
@@ -15,8 +15,8 @@ import pytest
 from data_canon.codebook.ctramp import CTRAMPStudentCategory
 from data_canon.codebook.persons import AgeCategory, Employment, SchoolType, Student
 from processing.formatting.ctramp.ctramp_config import CTRAMPConfig
-from processing.formatting.ctramp.mappings import (
-    ctramp_person_type_expression,
+from processing.formatting.ctramp.person_mappings import ctramp_person_type_expression
+from processing.formatting.ctramp.student_mappings import (
     ctramp_student_category_expression,
     log_student_category_warnings,
 )
